refactor(api): move update-certification-center-membership usecase
in team
